+++ packages/config/__tests__/Config.secure.test.ts
@@ -11,9 +11,11 @@
 
 import * as fs from "fs";
 import * as path from "path";
+import * as lodash from "lodash";
 import { CredentialManagerFactory } from "../../security";
 import { ImperativeError } from "../../error/src/ImperativeError";
 import { Config } from "../src/Config";
+import { IConfig } from "../src/doc/IConfig";
 import { IConfigSecure } from "../src/doc/IConfigSecure";
 import { IConfigVault } from "../src/doc/IConfigVault";
 
@@ -148,6 +150,118 @@ describe("Config secure tests", () => {
         expect(vault.load).toHaveBeenCalledTimes(1);
         expect(secureError).toBeDefined();
         expect(config.properties).toMatchSnapshot();
+    });
+
+    it("should list all secure fields in config layer", async () => {
+        jest.spyOn(Config, "search").mockReturnValue(projectConfigPath);
+        jest.spyOn(fs, "existsSync")
+            .mockReturnValueOnce(false)     // Project user layer
+            .mockReturnValueOnce(true)      // Project layer
+            .mockReturnValueOnce(false)     // User layer
+            .mockReturnValueOnce(false);    // Global layer
+        jest.spyOn(fs, "readFileSync");
+        const config = await Config.load(MY_APP);
+        expect(config.api.secure.secureFields()).toEqual(["profiles.fruit.properties.secret"]);
+    });
+
+    it("should list all secure fields for a profile", async () => {
+        jest.spyOn(Config, "search").mockReturnValue(projectConfigPath);
+        jest.spyOn(fs, "existsSync")
+            .mockReturnValueOnce(false)     // Project user layer
+            .mockReturnValueOnce(true)      // Project layer
+            .mockReturnValueOnce(false)     // User layer
+            .mockReturnValueOnce(false);    // Global layer
+        jest.spyOn(fs, "readFileSync");
+        const config = await Config.load(MY_APP);
+        expect(config.api.secure.securePropsForProfile("fruit.apple")).toEqual(["secret"]);
+    });
+
+    describe("secureInfoForProp", () => {
+        const configProperties: IConfig = {
+            profiles: {
+                fruit: {
+                    type: "fruit",
+                    profiles: {
+                        apple: {
+                            type: "apple",
+                            properties: {}
+                        }
+                    },
+                    properties: {}
+                }
+            },
+            defaults: {}
+        };
+
+        it("should return info for same level property when secure array includes property at higher level", () => {
+            const config = new (Config as any)();
+            const secureConfigProperties = lodash.cloneDeep(configProperties);
+            secureConfigProperties.profiles.fruit.secure = ["secret"];
+            jest.spyOn(config, "layerActive").mockReturnValueOnce({
+                exists: true,
+                properties: secureConfigProperties
+            });
+            expect(config.api.secure.secureInfoForProp("profiles.fruit.profiles.apple.properties.secret", false)).toMatchObject({
+                path: "profiles.fruit.profiles.apple.secure",
+                prop: "secret"
+            });
+        });
+
+        it("should return undefined when input is not a property path", () => {
+            const config = new (Config as any)();
+            expect(config.api.secure.secureInfoForProp("profiles.fruit")).toBeUndefined();
+        });
+
+        describe("when findUp is true", () => {
+            it("should return info for property when layer does not exist", () => {
+                const config = new (Config as any)();
+                jest.spyOn(config, "layerActive").mockReturnValueOnce({ exists: false });
+                expect(config.api.secure.secureInfoForProp("profiles.fruit.profiles.apple.properties.secret", true)).toMatchObject({
+                    path: "profiles.fruit.profiles.apple.secure",
+                    prop: "secret"
+                });
+            });
+
+            it("should return info for property when secure array does not exist", () => {
+                const config = new (Config as any)();
+                jest.spyOn(config, "layerActive").mockReturnValueOnce({
+                    exists: true,
+                    properties: configProperties
+                });
+                expect(config.api.secure.secureInfoForProp("profiles.fruit.profiles.apple.properties.secret", true)).toMatchObject({
+                    path: "profiles.fruit.profiles.apple.secure",
+                    prop: "secret"
+                });
+            });
+
+            it("should return info for same level property when secure array includes property at same level", () => {
+                const config = new (Config as any)();
+                const secureConfigProperties = lodash.cloneDeep(configProperties);
+                secureConfigProperties.profiles.fruit.profiles.apple.secure = ["secret"];
+                jest.spyOn(config, "layerActive").mockReturnValueOnce({
+                    exists: true,
+                    properties: secureConfigProperties
+                });
+                expect(config.api.secure.secureInfoForProp("profiles.fruit.profiles.apple.properties.secret", true)).toMatchObject({
+                    path: "profiles.fruit.profiles.apple.secure",
+                    prop: "secret"
+                });
+            });
+
+            it("should return info for higher level property when secure array includes property at higher level", () => {
+                const config = new (Config as any)();
+                const secureConfigProperties = lodash.cloneDeep(configProperties);
+                secureConfigProperties.profiles.fruit.secure = ["secret"];
+                jest.spyOn(config, "layerActive").mockReturnValueOnce({
+                    exists: true,
+                    properties: secureConfigProperties
+                });
+                expect(config.api.secure.secureInfoForProp("profiles.fruit.profiles.apple.properties.secret", true)).toMatchObject({
+                    path: "profiles.fruit.secure",
+                    prop: "secret"
+                });
+            });
+        });
     });
 
     describe("loadFailed", () => {